refactor: move types into related pages logic directory
