+++ src/pages/coinmarketcap.com/@types/scraped-project.ts
